add a degree raising function
There haven't been any actual tests of any of this and this is far out
enough that it might need such. Still, it's an algorithm from the paper,
albeit one of the easy ones. Multiplication and degree lowering are
likely the hard ones. Degree lowering is in effect reflecting detections
of the coefficients being in special position in the result type.
Multiplication, while it can be written out as some kind of isolated
operation easily enough, has what might be design questions alongside
the type hackery to be raised by it all in order to be able to use as
transparent and convenient an interface as possible.